+++ src/components/countdown/Main.tsx
@@ -2,13 +2,13 @@ import React, { Dispatch, SetStateAction, useEffect, useState } from "react";
 import { HStack, useToast } from "@chakra-ui/react";
 import { useTimer } from "react-timer-hook";
 import { useSound } from "use-sound";
-import sound from "../sounds/sound.mp3";
+import sound from "../../sounds/sound.mp3";
 import ActionSelectButton from "./ActionSelectButton";
 import DisplayTime from "./DisplayTime";
 import SettingTime from "./SettingTime";
 import MenuFunction from "./MenuFunction";
 import DisplaySettingTime from "./DisplaySettingTime";
-import useJudgeUpperLimitUnits from "../hooks/useJudgeUpperLimitUnits";
+import useJudgeUpperLimitUnits from "../../hooks/useJudgeUpperLimitUnits";
 
 type Props = {
   settingTime: number;
